Move composeTransactionSteps unit tests to correct directory
